Inside 'src,' create folder 'hooks,' which contains useProducts.js. This hook allows History and Input page to fetchProducts without repetitive code, promoting code reusability.
