+++ client/src/pages/History/index.js
@@ -1,9 +1,10 @@
-import React, { useEffect, useState, useCallback } from "react";
+import React, { useEffect, useState } from "react";
 import "./../../assets/fonts/fonts.css";
 import Navbar from "../../components/Navbar/Navbar";
 import Modal from "../../components/Modal/Modal";
 import ProductBar from "../../components/ProductBar/ProductBar";
 import IngredientModal from "../../components/IngredientModal/IngredientModal";
+import useProducts from "../../hooks/useProducts";
 import "./index.css";
 
 function History() {
@@ -13,7 +14,10 @@ function History() {
   const [productName, setProductName] = useState("");
   const [expirationDate, setExpirationDate] = useState("");
   const [productCategory, setProductCategory] = useState("");
-  const [products, setProducts] = useState([]);
+  // const [products, setProducts] = useState([]);
+
+  // Fetch products using the custom hook
+  const { products, fetchProducts } = useProducts();
 
   // Handles opening/closing the modal
   const toggleModal = () => {
@@ -29,31 +33,6 @@ function History() {
     }
     setIsIngredientModalOpen(!isIngredientModalOpen);
   };
-
-  const fetchProducts = useCallback(async () => {
-    const userId = sessionStorage.getItem("userId");
-    const apiUrl = `${process.env.REACT_APP_API_BASE_URL}/user/${userId}/get_products`;
-
-    try {
-      const response = await fetch(apiUrl);
-      const data = await response.json();
-      console.log("API Response:", data); // Check the structure of the response
-      if (response.ok) {
-        if (data) {
-          setProducts(data);
-        } else {
-          console.log("No products found.");
-        }
-      } else {
-        alert(`Error: ${data.error}`);
-      }
-    } catch (error) {
-      console.log("Error fetching products:", error);
-      alert(
-        "Failed to fetch products. Please try again later. Error: " + error
-      );
-    }
-  }, []);
 
   // Handles form submission
   const handleSubmit = async (event) => {
