+++ client/src/pages/InputForm/index.js
@@ -1,10 +1,11 @@
-import React, { useState, useCallback, useEffect } from "react";
+import React, { useState, useEffect } from "react";
 import "./../../assets/fonts/fonts.css";
 import Logo from "../../components/Logo/Logo";
 import { useNavigate } from "react-router-dom";
 import Navbar from "../../components/Navbar/Navbar";
 import Modal from "../../components/Modal/Modal";
 import IngredientModal from "../../components/IngredientModal/IngredientModal";
+import useProducts from "../../hooks/useProducts";
 import "./index.css";
 
 function InputForm() {
@@ -14,33 +15,11 @@ function InputForm() {
   const [productName, setProductName] = useState("");
   const [expirationDate, setExpirationDate] = useState("");
   const [productCategory, setProductCategory] = useState("");
-  const [products, setProducts] = useState([]);
+
+  // Fetch products using the custom hook
+  const { products, fetchProducts } = useProducts();
+
   const navigate = useNavigate();
-
-  const fetchProducts = useCallback(async () => {
-    const userId = sessionStorage.getItem("userId");
-    const apiUrl = `${process.env.REACT_APP_API_BASE_URL}/user/${userId}/get_products`;
-
-    try {
-      const response = await fetch(apiUrl);
-      const data = await response.json();
-      console.log("API Response:", data); // Check the structure of the response
-      if (response.ok) {
-        if (data) {
-          setProducts(data);
-        } else {
-          console.log("No products found.");
-        }
-      } else {
-        alert(`Error: ${data.error}`);
-      }
-    } catch (error) {
-      console.log("Error fetching products:", error);
-      alert(
-        "Failed to fetch products. Please try again later. Error: " + error
-      );
-    }
-  }, []);
 
   // Fetch products when the ingredient modal is opened
   useEffect(() => {
